% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_calc.R
\name{my_calc}
\alias{my_calc}
\title{Return result of a single parameter function.}
\usage{
my_calc(x, f)
}
\arguments{
\item{x}{Argument}

\item{f}{Function to be applied to x}
}
\value{
Result of the function
}
\description{
This is a demonstrator function, as part of the template.
}
\examples{
my_calc(1:10, mean)
}
